Moving octicon cache to lib
